Fix enforcer failure related to duplicated class.
Following insights by @ctrueden.
Conversation snippet:
- me:
"
I meet a enforcer-related compile error for this very project.
It depends on the bioformats_plugin artifact, and it generate the
following error:

[WARNING] Rule 3: org.apache.maven.plugins.enforcer.BanDuplicateClasses
failed with message:
No Duplicate Classes Allowed!
- For duplicate transitive dependencies, add dependency exclusions.
- For duplications between direct dependencies, resolve or add
  ignored classes to this rule's configuration.

  Found in:
    commons-io:commons-io:jar:2.6:compile
    ch.systems.cisd:jhdf5:jar:14.12.0:compile
  Duplicate classes:
    org/apache/commons/io/FileCleaningTracker.class
    org/apache/commons/io/comparator/SizeFileComparator.class
    org/apache/commons/io/output/ProxyWriter.class
"

- @ctrueden:
"
The problem is that ch.systems.cisd:jhdf5 is an uber-JAR.
You can avoid the error by adding:

<exclusions>
  <exclusion>
    <groupId>commons-io</groupId>
    <artifactId>commons-io</artifactId>
  </exclusion>
</exclusions>
in the <dependency> block of whatever dependency is pulling in
commons-io.
Best fix would be for us to deploy a new jhdf5 that does not do
this evilness.
I'd suggest checking the dependency tree using the Dependency
Hierarchy tab of the pom.xml in Eclipse.
You can search it for where the commons-io is coming in.
You need to plug all holes in your metaphorical boat.
"